\name{phenowheat}
\docType{data}
\alias{phenowheat}
\title{Durum wheat dataset}
\description{

This contain data from a multiparental durum wheat (Triticum turgidum L. spp. duram) trial consisting of 334 lines evaluated in a country-year combination. This population is characterized for Grain Yield (GY), grain volume weight (GVW), 1000-kernel weight (GWT) and heading date (HD) in the four environments. For further details see the vignettes in the package.

}

\usage{
  data(phenowheat)
}

\format{

  \enumerate{
    \item{\code{pheno}}{Matrix phenowheat.pheno contains the phenotypic data.}
    \item{\code{X}}{The matrix phenowheat.X contains the Genotypic data.}
  }

}


\keyword{datasets}
