\name{X}
\docType{data}
\alias{X}
\title{Durum Wheat X}
\description{

 Is a matrix (338 x 7594) with A balanced, four-way multiparental cross population was developed from four elite durum wheat cultivars (Neodur, Claudio,
Colosseo, and Rascon/Tarro) that were chosen as diverse contributors of different alleles of agronomic relevance.

}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\keyword{datasets}
