\name{pheno}
\alias{pheno}
\title{Durum Wheat}

\description{
This contain data from a multiparental durum wheat (Triticum turgidum L. spp. duram) trial consisting of 334 lines evaluated in a country-year combination, Arg12=Argentina 2012, Cad11=Canada 2011, Cad12=Canada 2012 y Pr11=Peru 2012. This population is characterized for Grain Yield (GY), grain volume weight (GVW), 1000-kernel weight (GWT) and heading date (HD) in the four environments.
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\keyword{datasets}
